[monorepo] feat: add seed job for monorepo layout
Add the seed job which can be used to automatically create the other package jobs.
The seed job can be configured to update the package jobs if changes are made to the buildConfiguration.yaml file.
The buildConfiguration.yaml has layout of each package and their dependencies.

fix: job properties getting remove

The ``with`` statement in groovy seem to create a new object which cause the previous properties to be lost.

fix: address feedback

Disable the ``UnnecessaryObjectReferences`` rule since unable to get groovy's ``with`` statement working in the Jenkins pipeline.