[mkdocs] fix: allow links in glossary definitions
The ezglossary plugin does not allow links (or any HTML tag) in glossary
definitions, because they need to go into a `name` attribute.
But instead of replacing links with their text, it just swallows them.
This commit patches the plugin's installation to keep the link text.

Also, change the plugin config and the template for slightly nicer HTML.