[mkdocs] feat: allow fast build through envvar
This was already controlled by a config setting, but it's more flexible this way.
When the SYMBOL_DOCS_DISABLE_TS environment variable exists and is set to true,
TypeScript docs are not generated. The docs will have missing links, but will build
much faster, which is useful for local development.