[mkdocs] fix: Name default JS classes properly
I don't know if it's Typedoc or Typescript, but default classes appear as a class named
"default" under a namespace with the class name.
This made keeping ref links in sync with Python very difficult, and the class is really
accessed using its name, so I just added a special case in the ref guide generation
for classes named "default".